refactor(LC007): split n-plus-one fixer helpers
- separate fix-context creation, syntax analysis, and fix application
- keep fixer behavior unchanged while reducing local complexity
